fix(web): stop prompting for provider updates
Remove passive provider update notifications from app startup. Keep provider update checks, settings actions, and sidebar progress for updates that users start themselves.
